Full SwiftUI conversion - removed storyboard and view controllers - Home renamed to MainView - MainView set as entry point - SwitUI views for MainView, row and rule entry - new Score file
