--- Fraxalator/RoundedCornersView.swift
@@ -1,21 +1,0 @@
-//
-//  RoundedCornersView.swift
-//  RWDevCon
-//
-//  Created by Mic Pringle on 26/02/2015.
-//  Copyright (c) 2015 Ray Wenderlich. All rights reserved.
-//
-
-import UIKit
-
-@IBDesignable
-class RoundedCornersView: UIView {
-  
-  @IBInspectable var cornerRadius: CGFloat = 0 {
-    didSet {
-      layer.cornerRadius = cornerRadius
-      layer.masksToBounds = cornerRadius > 0
-    }
-  }
-  
-}
