--- Fraxalator/SecondViewController.swift
@@ -1,25 +1,0 @@
-//
-//  SecondViewController.swift
-//  Fraxalator
-//
-//  Created by Nathan Mueller on 6/9/16.
-//  Copyright © 2016 OneNathan. All rights reserved.
-//
-
-import UIKit
-
-class SecondViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
