--- Fraxalator/TitleCell.swift
@@ -1,17 +1,0 @@
-//
-//  AnnotatedPhotoCell.swift
-//  Opio-Assist
-//
-//  Created by Nathan Mueller on 5/15/16.
-//  Copyright © 2016 OneNathan. All rights reserved.
-//
-
-import UIKit
-
-class TitleCell: UICollectionViewCell {
-    
-
-    @IBOutlet weak var title: UILabel!
-    var cellLabel : String!
-    
-}
